pyproject.toml
[tool.poetry]
name = "oil"
version = "0.1.0"
description = "API to fetch and create oil resources"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13.3"
fastapi = "^0.115.12"
uvicorn = "^0.32.0"
sqlmodel = "^0.0.24"
asyncpg = "^0.30.0"
alembic = "^1.13.2"
tenacity = "^8.3.0"
pydantic = "^2.7.1"
email-validator = "^2.0.0"
pydantic-settings = "^2.2.1"
psycopg2-binary = "^2.9.9"
greenlet = "^3.2.1"
boto3 = "1.35.99"
fastapi-pagination = "^0.12.34"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}  # JWT token handling
httpx = "^0.27.0"  # Async HTTP client for JWKS endpoint
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
flake8 = "*"
mypy = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.13"
disallow_untyped_defs = true
disallow_incomplete_defs = true
